% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggpixel}
\alias{ggpixel}
\title{Plot a word on tiles}
\usage{
ggpixel(x, noise.difference = 5, color.palette = c("viridis", "magma",
  "plasma", "inferno"), margin.left = 5, margin.right = 5,
  margin.top = 10, margin.bottom = 10, letter.spacing = 3,
  overlay = NULL, overlay.radius = 20, overlay.color = "white",
  na.value = "white", color = "grey30", aspect.ratio = 0.1)
}
\arguments{
\item{x}{a \code{character} vector}

\item{noise.difference}{\code{numeric} value that is added to the letter
tiles in order to create a contrast. The higher the stronger the contrast.}

\item{color.palette}{obviously the colorpalette.
One of the viridis colors.}

\item{margin.left}{how many pixels are inserted to the left from the string}

\item{margin.right}{how many pixels are inserted to the right
from the string}

\item{margin.top}{how many pixels are inserted to the top  from the string}

\item{margin.bottom}{how many pixels are inserted to the bottom
from the string}

\item{letter.spacing}{how many pixels are used for spacing two characters}

\item{overlay}{how many corners the overlay will have. Per default (NULL)
no overlay is used}

\item{overlay.radius}{how many pixels the radius of the overlay will be}

\item{overlay.color}{the color of the overlay}

\item{na.value}{color used for NAs}

\item{color}{color used for the lines separating the tiles}

\item{aspect.ratio}{\code{ggplot2} \code{aspect.ratio} parameter. Modifies
the dimensions of the tiles.}
}
\value{
returns a \code{ggplot} object
}
\description{
The method takes a string and translates it into a tile plot
 using \code{ggplot2}. The resulting lpot looks like a heamtap with a string
 engraved.
}
\examples{
\dontrun{
  obiwan <- ggpixel("hello there!")

  grievous <- ggpixel("general kenobi!", 3, "magma", 10, 10, 10, 10)
}
}
